refac: devide app/gateway/main.go #5
